# WareHouseProjects/api/docs/swagger.yaml
definitions:
...
  models.ComingTableProduct:
    properties:
      barcode:
        type: string
      category_id:
        type: string
      coming_table_id:
        type: string
      count:
        type: number
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      total_price:
        type: number
      updated_at:
        type: string
    type: object
...
  models.CreateComingTableProductSwagger:
    properties:
      barcode:
        type: string
      coming_table_id:
        type: string
      count:
        type: number
    type: object
...
  models.GetAllComingTableProductRequest:
    properties:
      barcode:
        type: string
      category_id:
        type: string
      limit:
        type: integer
      page:
        type: integer
    type: object
...
  models.UpdateComingTableProduct:
    properties:
      barcode:
        type: string
      category_id:
        type: string
      coming_table_id:
        type: string
      count:
        type: number
      id:
        type: string
      name:
        type: string
      price:
        type: number
      total_price:
        type: number
    type: object
...
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /coming_table_product:
    get:
      consumes:
      - application/json
      description: gets all Coming_TableProduct based on limit, page and search by
        name
      parameters:
      - default: 10
        description: limit
        in: query
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: category_id
        in: query
        name: category_id
        type: string
      - description: barcode
        in: query
        name: barcode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllComingTableProductRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: LIST Coming_Table_Product
      tags:
      - coming_table_product
    post:
      consumes:
      - application/json
      description: add ComingTableProduct data to db based on given info in body
      parameters:
      - description: ComingTableProduct data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateComingTableProductSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: CREATE ComingTableProduct
      tags:
      - coming_table_product
  /coming_table_product/{id}:
    delete:
      consumes:
      - application/json
      description: deletes ComingTableProduct by id
      parameters:
      - description: id of ComingTableProduct
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: DELETE ComingTableProduct BY ID
      tags:
      - coming_table_product
    get:
      consumes:
      - application/json
      description: gets ComingTableProduct by ID
      parameters:
      - description: ComingTableProduct ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ComingTableProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET BY ID
      tags:
      - coming_table_product
    put:
      consumes:
      - application/json
      description: UPDATES COMINGTableProduct BASED ON GIVEN DATA AND ID
      parameters:
      - description: id of ComingTableProduct
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: ComingTableProduct data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateComingTableProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: UPDATE COMINGTableProduct
      tags:
      - coming_table_product
...
swagger: "2.0"